rename server dir to logic
